refactor: add componets for calorie and muscle badges
